Rename all easy_transformer instances to transformer_lens
