+++ easy_transformer/__init__.py
@@ -1,10 +1,16 @@
-import sys
+"""Alias for TransformerLens
 
-# make sure bar is in sys.modules
+EasyTransformer has been renamed to TransformerLens. This is an alias package so
+that the change is non-breaking. However you should instead import
+`transformer_lens`."""
+import sys
+from warnings import warn
+
+# Make sure transformer_lens is in sys.modules
 import transformer_lens
 
-# link this module to bar
-sys.modules[__name__] = sys.modules['transformer_lens']
+# Warn the user that this has been renamed
+warn("easy_transformer has been renamed to transformer_lens")
 
-# Or simply
-# sys.modules[__name__] = __import__('transformer_lens')
+# Link this module to transformer_lens
+sys.modules[__name__] = sys.modules['transformer_lens']
