+++ transformer_lens/__init__.py
@@ -2,13 +2,14 @@ from . import hook_points
 from . import utils
 from . import evals
 from .past_key_value_caching import (
-    EasyTransformerKeyValueCache,
-    EasyTransformerKeyValueCacheEntry,
+    TransformerLensKeyValueCache,
+    TransformerLensKeyValueCacheEntry,
 )
 from . import components
-from .EasyTransformerConfig import EasyTransformerConfig
+from .TransformerLensConfig import TransformerLensConfig
 from .FactoredMatrix import FactoredMatrix
 from .ActivationCache import ActivationCache
-from .EasyTransformer import EasyTransformer
+from .TransformerLens import TransformerLens
+from .TransformerLens import TransformerLens
 from . import loading_from_pretrained as loading
 from . import train
